refactor: extract quality HTTP routes
